[#1964] Move safePropertyExists to utils, de-duplicate ability logic
